Added function to return all actions, refactored structure
