+++ lambda/actions/dynamo_delete_table.py
