+++ lambda/actions/ec2_delete_internet_gateway.py
