+++ lambda/actions/ec2_delete_key_pair.py
