+++ lambda/actions/ec2_delete_nat_gateway.py
