+++ lambda/actions/ec2_delete_security_group.py
