+++ lambda/actions/ec2_delete_snapshot.py
