+++ lambda/actions/ec2_snapshot_instance.py
