+++ lambda/actions/ec2_stop_instance.py
