+++ lambda/actions/ec2_terminate_instance.py
