+++ lambda/actions/elb_delete_load_balancer.py
