+++ lambda/actions/redshift_delete_cluster.py
