+++ lambda/actions/workspaces_stop_workspace.py
